requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.7.1

# API
fastapi==0.111.0
uvicorn==0.30.1

# Data handling
pandas==2.2.2
numpy==1.26.4

# LLM (Groq)
groq==0.9.0

# Utilities
loguru==0.7.2
tqdm==4.66.4

# File handling
python-multipart==0.0.9

# Frontend
streamlit==1.35.0

# JSON handling
orjson==3.10.3

# Optional
tabulate==0.9.0